+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 
 - [#366] fix native libraries sometimes breaking when already exists
 - Fix old v1 instances not converting their format when necessary
+- When updating, mark the launcher file as executable